Finished the Buddy script
It will follow the players jump etc, move within the distance you set to
the player (target) and when you get to a certain distance to the lake
transform, that you set, the buddy will head off to the lake and then
disappear when it gets to within 0.5 of the lake transform.  If you need
it to run faster to the lake, then include a multiplyer in that
statement so that it goes faster.